pyproject.toml
[project]
name = "ccsc-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "django-browser-reload>=1.21.0",
    "django-htmx>=1.27.0",
    "django-unfold>=0.80.2",
    "django-vite>=3.1.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]
